+++ src/wolframite/base/cep.clj
@@ -15,11 +15,16 @@
   Evaluate: the Expr on (some) Wolfram Engine
   Parse:    returned result into clj data.
   Each stage can be skipped with appropriate `opts` `:flag` e.g. `:no-parse`"
-  [expr {:keys [flags]
-         :as   opts}]
-  (let [convert (if (options/flag?' flags flags/convert) convert/convert identity-first)
-        evaluate (if (options/flag?' flags flags/evaluate) evaluate/evaluate identity-first)
-        parse (if (options/flag?' flags flags/parse) parse/parse identity-first)
+  [expr {:keys [flags] :as opts}]
+  (let [convert (if (options/flag?' flags flags/convert)
+                  convert/wrapping-convert
+                  identity-first)
+        evaluate (if (options/flag?' flags flags/evaluate)
+                   evaluate/evaluate
+                   identity-first)
+        parse (if (options/flag?' flags flags/parse)
+                parse/unwrapping-parse
+                identity-first)
         res (-> expr
                 (convert opts)
                 (evaluate opts)