+++ src/wolframite/impl/error_detection.clj
@@ -1,12 +1,8 @@
 (ns wolframite.impl.error-detection
   "Centralize Wolfram-related error handling"
   (:require [clojure.tools.logging :as log]
+            [wolframite.impl.internal-constants :as internal-constants]
             [wolframite.impl.protocols :as proto]))
-
-;; Wolfram sometimes indicates failure by returning the symbol $Failed, at least in some cases
-(defonce ^:private failed-expr-p
-         ;; Delay b/c we need to wait until JLink is loaded
-         (delay (proto/atomic-expr proto/type-symbol "$Failed")))
 
 (defn error-message-expr
   "Does the result represent a (delayed) error message template, which we can evaluate
@@ -20,12 +16,21 @@
          (when (-> MessageTemplate meta :wolfram/delayed)
            MessageTemplate))))
 
+(defn- unchanged-expression? [input output]
+  (let [size-wrapper? (= (proto/head-sym-str input)
+                         (name internal-constants/wolframiteLimitSize))
+        unwrapped-input (if size-wrapper?
+                          (-> input proto/args first)
+                          input)]
+    (= unwrapped-input output)))
+
 (defn ensure-no-eval-error [expr eval-result eval-messages]
-  (let [messages-text (mapv :content eval-messages)]
+  (let [messages-text (mapv :content eval-messages)
+        failed-sym? (proto/failed? eval-result)]
     (cond
       (and (seq eval-messages)
-           (or (= eval-result @failed-expr-p)
-               (= eval-result expr)))
+           (or failed-sym?
+               (unchanged-expression? expr eval-result)))
       ;; If input expr == output expr, this usually means the evaluation failed
       ;; (or there was nothing to do); if there are also any extra text/message packets
       ;; then it most likely has failed, and those messages explain what was wrong
@@ -39,7 +44,7 @@
                        :messages eval-messages
                        :result eval-result}))
 
-      (= eval-result @failed-expr-p) ; but no messages
+      failed-sym? ; but no messages
       (throw (ex-info (str "Evaluation has failed. Result: "
                            eval-result
                            " No details available.")