+++ test/wolframite/base/parse_test.clj
@@ -11,39 +11,39 @@
   (#'wl/init-jlink! (deref #'wl/kernel-link-atom) {}) ; bypass private var via deref #'
   (testing "atomic expr"
     (is (= "str"
-           (parse/parse (convert/convert "str" nil) nil))
+           (parse/unwrapping-parse (convert/convert "str" nil) nil))
         "string")
     (is (= 1
-           (parse/parse (convert/convert 1 nil) nil))
+           (parse/unwrapping-parse (convert/convert 1 nil) nil))
         "int")
     (is (= Long/MAX_VALUE
-           (parse/parse (convert/convert Long/MAX_VALUE nil) nil))
+           (parse/unwrapping-parse (convert/convert Long/MAX_VALUE nil) nil))
         "long")
     (is (= 1.1
-           (parse/parse (convert/convert 1.1 nil) nil))
+           (parse/unwrapping-parse (convert/convert 1.1 nil) nil))
         "float")
     (is (= (bigint 333)
-           (parse/parse (convert/convert (bigint 333) nil) nil))
+           (parse/unwrapping-parse (convert/convert (bigint 333) nil) nil))
         "bigint")
     #_ ; Not sure how to test w/o actually calling eval...
     (is (= (/ 1 3)
-           (parse/parse (convert/convert (/ 1 3) nil) nil))
+           (parse/unwrapping-parse (convert/convert (/ 1 3) nil) nil))
         "ratio")
     (is (= 'MySym
-           (parse/parse (convert/convert 'MySym nil) nil))
+           (parse/unwrapping-parse (convert/convert 'MySym nil) nil))
         "symbol"))
   (testing "maps"
     (is (= {"a" 1}
-           (parse/parse (convert/convert {"a" 1} nil) nil)))
+           (parse/unwrapping-parse (convert/convert {"a" 1} nil) nil)))
     (is (= {}
-           (parse/parse (convert/convert {} nil) nil))
+           (parse/unwrapping-parse (convert/convert {} nil) nil))
         "Empty maps work too"))
   (testing "RuleDelayed"
     ;; Some errors are returned as a MessageTemplate -> delayed call to get the message name, f.ex.
     ;; `(wl/! (list (w/Interpreter "City") "San Francisco, US"))` while offline.
     (let [parsed (-> (w/Association (w/RuleDelayed "MessageTemplate" (w/MessageName w/Interpreter "noknow")))
                      (convert/convert nil)
-                     (parse/parse nil))]
+                     (parse/unwrapping-parse nil))]
       (is (= {"MessageTemplate" '(MessageName Interpreter "noknow")}
              parsed)
           "We can parse RuleDelayed")
@@ -73,30 +73,30 @@
       (testing "w/ named args"
         (is (= '(fn [x] x)
                (-> (convert/convert (convert/->wolfram-str-expr "Function[{x},x]") ctx)
-                   (parse/parse nil))))
+                   (parse/unwrapping-parse nil))))
         (is (= '(fn [x y] (+ x y))
                (-> (convert/convert (convert/->wolfram-str-expr "Function[{x,y},x + y]") ctx)
-                   (parse/parse nil)))))
+                   (parse/unwrapping-parse nil)))))
       (testing "w/ anonymous args"
         ;; Fn w/ named args: Function[{x}, Plus[x, 1]]
         ;; Lambda w/ anonymous args: Function[Plus[(Slot 1), 1]]
         (is (= '(fn (Slot 1))
                (-> (convert/convert (convert/->wolfram-str-expr "#&") ctx)
-                   (parse/parse nil)))
+                   (parse/unwrapping-parse nil)))
             "An anonymous, one-arg lambda becomes Function (aliased to fn) w/ Slot
             (# = #1 = Slot 1; `body&` is the fn)")
         (is (= '(fn (+ (Slot 1) (Slot 2)))
                (-> (convert/convert (convert/->wolfram-str-expr "#1 + #2&") ctx)
-                   (parse/parse nil))))))))
+                   (parse/unwrapping-parse nil))))))))
 
 (deftest fix-minus-reverse-aliased
   (wl/start!) ; we need the Kernel for turning stringified Wolfram into Expr
   (let [ctx {:jlink-instance (jlink-instance/get)}]
     (is (= '(- 42)
            (-> (convert/convert (convert/->wolfram-str-expr "Minus[42]") ctx)
-               (parse/parse nil)))
+               (parse/unwrapping-parse nil)))
         "Minus is parsed as the alias `-`")
    (is (= '(- 4 2)
           (-> (convert/convert (convert/->wolfram-str-expr "Subtract[4, 2]") ctx)
-              (parse/parse nil)))
+              (parse/unwrapping-parse nil)))
        "Subtract is parsed as the alias `-`")))